+++ db/sqlc/transfer.sql.go
@@ -179,7 +179,7 @@ func (q *Queries) ListTransfersByToAccount(ctx context.Context, arg ListTransfer
 }
 
 const totalAmountReceivedByAccount = `-- name: TotalAmountReceivedByAccount :one
-SELECT COALESCE(SUM(amount), 0) AS total_received
+SELECT COALESCE(SUM(amount)::BIGINT, 0) AS total_received
 FROM transfers
 WHERE to_account_id = $1
 `
@@ -192,7 +192,7 @@ func (q *Queries) TotalAmountReceivedByAccount(ctx context.Context, toAccountID 
 }
 
 const totalAmountSentByAccount = `-- name: TotalAmountSentByAccount :one
-SELECT COALESCE(SUM(amount), 0) AS total_sent
+SELECT COALESCE(SUM(amount)::BIGINT, 0) AS total_sent
 FROM transfers
 WHERE from_account_id = $1
 `
